update: authorization implemented

+++ src/api/notes/handler.js
@@ -1,3 +1,5 @@
+const ClientError = require("../../exceptions/ClientError");
+
 class NotesHandler {
   constructor(service, validator) {
     this._service = service;
@@ -15,10 +17,16 @@ class NotesHandler {
 
   async postNoteHandler(request, h) {
     this._validator.validateNotePayload(request.payload);
+    const { id: credentialId } = request.auth.credentials;
     const { title = "untitled", body, tags } = request.payload;
 
-    const noteId = await this._service.addNote({ title, body, tags });
-
+    const noteId = await this._service.addNote({
+      title,
+      body,
+      tags,
+      owner: credentialId,
+    });
+    // Server ERROR!
     const response = h.response({
       status: "success",
       message: "Catatan berhasil ditambahkan",
@@ -30,8 +38,9 @@ class NotesHandler {
     return response;
   }
 
-  async getNotesHandler() {
-    const notes = await this._service.getNotes();
+  async getNotesHandler(request) {
+    const { id: credentialId } = request.auth.credentials;
+    const notes = await this._service.getNotes(credentialId);
     return {
       status: "success",
       data: {
@@ -42,7 +51,8 @@ class NotesHandler {
 
   async getNoteByIdHandler(request, h) {
     const { id } = request.params;
-    const note = await this._service.getNoteById(id);
+    const { id: credentialId } = request.auth.credentials;
+    const note = await this._service.getNoteById(id, credentialId);
     return {
       status: "success",
       data: {
@@ -54,7 +64,9 @@ class NotesHandler {
   async putNoteByIdHandler(request, h) {
     this._validator.validateNotePayload(request.payload);
     const { id } = request.params;
+    const { id: credentialId } = request.auth.credentials;
 
+    await this._service.verifyNoteOwner(id, credentialId);
     await this._service.editNoteById(id, request.payload);
 
     const updatedNote = await this._service.getNoteById(id);
@@ -69,13 +81,35 @@ class NotesHandler {
   }
 
   async deleteNoteByIdHandler(request, h) {
-    const { id } = request.params;
-    await this._service.deleteNoteById(id);
+    try {
+      const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+      await this._service.verifyNoteOwner(id, credentialId);
+      await this._service.deleteNoteById(id);
 
-    return {
-      status: "success",
-      message: "Catatan berhasil dihapus",
-    };
+      return {
+        status: "success",
+        message: "Catatan berhasil dihapus",
+      };
+    } catch (error) {
+      if (error instanceof ClientError) {
+        const response = h.response({
+          status: "fail",
+          message: error.message,
+        });
+        response.code(error.statusCode);
+        return response;
+      }
+
+      // Server ERROR!
+      const response = h.response({
+        status: "error",
+        message: "Maaf, terjadi kegagalan pada server kami.",
+      });
+      response.code(500);
+      console.error(error);
+      return response;
+    }
   }
 }
